Se modularizo auth y se acomodaron los demas modulos devido a este cambio
