+++ gym-back-end/src/app.module.ts
@@ -2,8 +2,22 @@ import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { CommonModule } from './common/common.module';
+import { AuthModule } from './auth/auth.module';
+import { TypeOrmModule } from '@nestjs/typeorm';
 @Module({
-  imports: [CommonModule], 
+  imports: [
+    TypeOrmModule.forRoot({
+      type: 'mysql',
+      host: process.env.DB_HOST,
+      port: Number(process.env.DB_PORT),
+      username: process.env.DB_USER,
+      password: process.env.DB_PASSWORD,
+      database: process.env.DB_DATABASE,
+      synchronize: false,
+      entities: ['dist/**/*.entity.js'],
+      // logging: 'all',
+  }),
+    CommonModule, AuthModule], 
   controllers: [AppController],
   providers: [AppService],
 })
