+++ gym-back-end/src/auth/controllers/login/login.controller.ts
@@ -1,7 +1,7 @@
 import { Body, Controller, Get, HttpException, HttpStatus, InternalServerErrorException, Post, Req, Res, UnauthorizedException } from '@nestjs/common';
 import { Request, Response } from 'express';
-import { LoginService } from '../services/login.service';
-import { RegisterService } from '../services/register.service';
+import { LoginService } from '../../services/login/login.service';
+import { RegisterService } from '../../services/register/register.service';
 import { jwtDecode } from 'jwt-decode';
 import { IUserResponse } from 'src/common/models/interfaces/user.interface';
