+++ gym-back-end/src/common/common.module.ts
@@ -1,9 +1,5 @@
 import { Module } from '@nestjs/common';
-import { RegisterService } from 'src/auth/services/register.service';
-import { RegisterController } from '../auth/controllers/register.controller';
 import { JwtModule } from '@nestjs/jwt';
-import { LoginService } from '../auth/services/login.service';
-import { LoginController } from '../auth/controllers/login.controller';
 import * as dotenv from 'dotenv';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from './models/entities/user.entity';
@@ -22,25 +18,10 @@ if (result.error) {
 
 @Module({
     imports: [
-        TypeOrmModule.forRoot({
-            type: 'mysql',
-            host: process.env.DB_HOST,
-            port: Number(process.env.DB_PORT),
-            username: process.env.DB_USER,
-            password: process.env.DB_PASSWORD,
-            database: process.env.DB_DATABASE,
-            synchronize: false,
-            entities: ['dist/**/*.entity.js'],
-            // logging: 'all',
-        }),
-        TypeOrmModule.forFeature([User, Appointment, Booking]),
-        JwtModule.register({
-            secret: process.env.JWT_SECRET_KEY,
-            signOptions: { expiresIn: '60s' },
-        })
+        TypeOrmModule.forFeature([Appointment, Booking]),
     ],
-    controllers: [RegisterController, LoginController, AppointmentController, BookingController],
-    providers: [RegisterService, LoginService, AppointmentService, BookingService],
+    controllers: [AppointmentController, BookingController],
+    providers: [AppointmentService, BookingService],
     exports: [],
 })
 export class CommonModule { }
